+++ src/game-core/domain/index.ts
@@ -1,0 +1,9 @@
+export type * from "./actions";
+export type * from "./events";
+export type * from "./factions";
+export type * from "./ideology";
+export type * from "./institutions";
+export type * from "./primitives";
+export type * from "./relations";
+export type * from "./results";
+export type * from "./state";